文字起こし/requirements.txt
flask
openai-whisper
python-docx
torch
